+++ agent/impl/src/main/java/com/walmartlabs/concord/agent/ExecutionManager.java
@@ -2,8 +2,8 @@ package com.walmartlabs.concord.agent;
 
 import com.google.common.cache.Cache;
 import com.google.common.cache.CacheBuilder;
-import com.walmartlabs.concord.common.DependencyManager;
 import com.walmartlabs.concord.common.IOUtils;
+import com.walmartlabs.concord.dependencymanager.DependencyManager;
 import com.walmartlabs.concord.rpc.AgentApiClient;
 import com.walmartlabs.concord.rpc.JobStatus;
 import com.walmartlabs.concord.rpc.JobType;